updating vscode setting
That said, I'm under the impression the vscode devs are interested in
using foss editor plugins through LSP, but are not interested in making
their extensions available through LSP. I'm very unhappy with that - it
undermines the entire point of LSP in my eyes. So I stop using vscode
for now.